Add discord alerts for crashes and doc suggestions
Adds some error detection and tries to notify Discord when something
explodes. Error detection ignores rate limiting responses from Coda, any
rate limiting from Coda just results in the script dying, which
naturally leads to a few minute wait for the next run.

Also adds Discord notifications for suggestions to answers

Both of the above are sent via webhooks, which are kept secret for
hopefully understandable reasons. If you'd like to run this thing
yourself you will need to request a channel webhook for the test server
or use a webhook to a test channel on your own server, either is fine.